my-ai-app/requirements.txt
fastapi
uvicorn
torch
transformers
diffusers
accelerate
pydub
torchaudio
barked

